+++ tests/test_basic_auth_integration.py
@@ -1,7 +1,7 @@
 from dash import Dash, Input, Output, dcc, html
 import requests
 
-from dash_auth import basic_auth, add_public_routes
+from dash_auth import BasicAuth, add_public_routes, protected
 
 
 TEST_USERS = {
@@ -26,7 +26,7 @@ def test_ba001_basic_auth_login_flow(dash_br, dash_thread_server):
     def update_output(new_value):
         return new_value
 
-    basic_auth.BasicAuth(app, TEST_USERS["valid"], public_routes=["/home"])
+    BasicAuth(app, TEST_USERS["valid"], public_routes=["/home"])
     add_public_routes(app, ["/user/<user_id>/public"])
 
     dash_thread_server(app)
@@ -58,3 +58,47 @@ def test_ba001_basic_auth_login_flow(dash_br, dash_thread_server):
         # visiting the page again will use the saved credentials
         dash_br.driver.get(base_url)
         dash_br.wait_for_text_to_equal("#output", "initial value")
+
+
+def test_ba002_basic_auth_groups(dash_br, dash_thread_server):
+    app = Dash(__name__)
+    app.layout = html.Div([
+        dcc.Input(id="input", value="initial value"),
+        html.Div(id="output")
+    ])
+
+    @app.callback(
+        Output("output", "children"),
+        Input("input", "value"),
+        groups=["admin"],
+    )
+    @protected(
+        unauthenticated_output="unauthenticated",
+        missing_permissions_output="forbidden",
+        groups=["admin"],
+    )
+    def update_output(new_value):
+        return new_value
+
+    BasicAuth(
+        app,
+        TEST_USERS["valid"],
+        public_routes=["/home"],
+        user_groups={"hello": ["admin"]},
+        secret_key="Test!",
+    )
+
+    dash_thread_server(app)
+    base_url = dash_thread_server.url
+
+    for user, password in TEST_USERS["valid"]:
+        # login using the URL instead of the alert popup
+        # selenium has no way of accessing the alert popup
+        dash_br.driver.get(base_url.replace("//", f"//{user}:{password}@"))
+
+        # the username:password@host url doesn"t work right now for dash
+        # routes, but it saves the credentials as part of the browser.
+        # visiting the page again will use the saved credentials
+        dash_br.driver.get(base_url)
+        expected = "initial value" if user == "hello" else "forbidden"
+        dash_br.wait_for_text_to_equal("#output", expected)