+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- OIDCAuth allows to authenticate via OIDC
+- BasicAuth saves the current user in the session
+- Ability to define user groups in BasicAuth
+- Group-based permission and protection functions
+
 ## [2.2.0] - 2024-02-05
 ### Added
 - Possibility to whitelist routes with the `add_public_routes` utility function, the routes should follow Flask route syntax